conway_gol_test.py
#!/usr/bin/env python
"""Tests for Game of Life.
"""
import unittest
import textwrap
from conway_gol import ConwayGameOfLife


class TestCell(unittest.TestCase):

    def setUp(self):
        self.gol = ConwayGameOfLife()

    def test_get_neighbour_cells(self):
        """Test get_neighbour_cells.
        """
        expected = self.gol.to_cells(
            (-1, 1), (0, 1), (1, 1), (1, 0),
            (1, -1), (0, -1), (-1, -1), (-1, -0)
        )
        self.assertEqual(
            self.gol._get_neighbour_cells(
                self.gol.Cell(0, 0)), expected)

    def test_get_next_world(self):
        """Test get_next_world.
        """
        world = self.gol.to_cells((0, 0), (1, 1), (0, 1))
        # TODO check manually...
        expected = self.gol.to_cells((0, 1), (1, 0), (0, 0), (1, 1))
        self.assertEqual(self.gol._get_next_world(world), expected)

    def test_generator(self):
        """Test conway generator.
        """
        world = self.gol.to_cells((0, 0), (1, 1), (0, 1), (1, 2), (1, 0))
        gol_gen = self.gol.conway_gol(initial_state=world)
        # Run for 3 generations
        for _ in xrange(3):
            gol_gen.next()
        # TODO check manually...
        # check 4th generation
        expected = self.gol.to_cells((1, 1), (2, 1))
        self.assertEqual(gol_gen.next(), expected)

    def test_asciitext_to_template(self):
        """Test asciitext_to_template.
        """
        text = "ooo\noZo\nooo"
        expected = [(-1, -1), (0, -1), (1, -1),
                    (-1, 0), (1, 0),
                    (-1, 1), (0, 1), (1, 1)]
        self.assertEqual(
            self.gol.asciitext_to_template(
                text, 'o', 'Z', False), expected)

    # Test Wikipedia examples

    def test_diehard(self):
        """..."Diehard" is a pattern that eventually disappears
              (rather than merely stabilizing) after 130 generations...
        """
        diehard = textwrap.dedent(
            """\
            ..........
            .......x..
            .xx.......
            ..x...xxx.
            ..........
            """)

        world = self.gol.to_cells(
            *self.gol.asciitext_to_template(diehard))
        gol_gen = self.gol.conway_gol(initial_state=world)
        # Run for 128 generations
        for _ in xrange(129):
            gol_gen.next()
        # check 129th generation is still alive
        self.assertTrue(gol_gen.next())
        # check 130th generation is dead
        self.assertFalse(gol_gen.next())

    def test_acorn(self):
        """..."Acorn" takes 5206 generations to generate 633 cells
              including 13 escaped gliders...
        """
        acorn = textwrap.dedent(
            """\
            .........
            ..x......
            ....x....
            .xx..xxx.
            .........
            """)

        world = self.gol.to_cells(
            *self.gol.asciitext_to_template(acorn))
        gol_gen = self.gol.conway_gol(initial_state=world)
        # Run for 5206 generations
        for _ in xrange(5206):
            gol_gen.next()
        # off-by-one? Not sure why...
        self.assertNotEqual(gol_gen.send(True), 633)
        gol_gen.next()  # 5207th generation
        # check that there are 633 cells
        self.assertEqual(gol_gen.send(True), 633)

    def test_pulsar(self):
        """...The "pulsar" is the most common period 3 oscillator....
        """
        pulsar = textwrap.dedent(
            """\
            ...............
            ...xxx...xxx...
            ...............
            .x....x.x....x.
            .x....x.x....x.
            .x....x.x....x.
            ...xxx...xxx...
            ...............
            ...xxx...xxx...
            .x....x.x....x.
            .x....x.x....x.
            .x....x.x....x.
            ...............
            ...xxx...xxx...
            ...............
            """)

        world = self.gol.to_cells(
            *self.gol.asciitext_to_template(pulsar))
        gol_gen = self.gol.conway_gol(initial_state=world)
        expected = (gol_gen.next(), gol_gen.next(), gol_gen.next())
        self.assertEqual(
            (gol_gen.next(), gol_gen.next(), gol_gen.next()), expected)


if __name__ == '__main__':
    unittest.main()
